# mcp-service/openapi.yaml
openapi: 3.0.0
info:
  title: Restaurant Management API
  description: API for managing Indian restaurants, menus, and orders
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
  /restaurants:
    get:
      operationId: getRestaurants
      summary: Get all restaurants
      description: Returns a list of all Indian restaurants
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
    
    post:
      operationId: createRestaurant
      summary: Create a new restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantInput'
      responses:
        '201':
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
...
  /orders:
    get:
      operationId: getOrders
      summary: Get all orders
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    
    post:
      operationId: createOrder
      summary: Create a new order
      description: Create order with automatic GST calculation (5%)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
...
components:
  schemas:
    Restaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone_number:
          type: string
        cuisine_type:
          type: string
        created_at:
          type: string
          format: date-time
    RestaurantInput:
      type: object
      required:
        - name
        - address
      properties:
        name:
          type: string
        address:
          type: string
        phone_number:
          type: string
        cuisine_type:
          type: string
          default: "Indian"
...
    Order:
      type: object
      properties:
        id:
          type: integer
        restaurant_id:
          type: integer
        customer_name:
          type: string
        customer_phone:
          type: string
        order_items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status:
          type: string
        total_amount:
          type: number
        tax_amount:
          type: number
        discount:
          type: number
        final_amount:
          type: number
        payment_status:
          type: string
        payment_method:
          type: string
        billing_address:
          type: string
    OrderInput:
      type: object
      required:
        - restaurant_id
        - customer_name
        - order_items
      properties:
        restaurant_id:
          type: integer
        customer_name:
          type: string
        customer_phone:
          type: string
        order_items:
          type: array
          items:
            type: object
            properties:
              menu_item_id:
                type: integer
              quantity:
                type: integer
              price:
                type: number
              notes:
                type: string
        discount:
          type: number
          default: 0
        payment_method:
          type: string
          enum: [cash, card, upi, digital_wallet]
        billing_address:
          type: string
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        menu_item_id:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        notes:
          type: string
        subtotal:
          type: number